requirements.txt
llama-cpp-python 
pyttsx3 
speechrecognition 
pyaudio
